attrs_ui: add AttrsWidget
This (and its tests) is copied from the uawidgets project and lightly
updated to support the async API from asyncua. It's not a complete port:
there are a few widgets from uawidgets that we're not including here
that prevent the editing of data types, but we can revisit this later if
we need it.

Signed-off-by: Kyle Fazzari <kyle@miriamtech.com>